Add concurrency limit for DVC routes
